Extract point cloud transformers into separate files
